Organized files, edited menu style (#31)
Created a new file for the menu component and moved it along with the `sketchfabIntegration,.js` into the `src` folder. 
This is to help keep the main `index..js` file neat and organized. This will also help keep the repo easy to work on as more items and features are added.